Eemalda legacy prosopography _legacy_ops moodul (#73)
_legacy_ops.py (2012 rida) kustutatud. Domeeniloogika elab nüüd otse
person_crud / person_search / relations / indices / merge_ops moodulites;
ops.py jääb ühilduvusfassaadiks. Kollektsiooni/haridusasutuse membership-
loogika kolis indices.py-sse muutmata kujul.

_compat.py sünkroniseerib monkeypatch'id otse state'i ja laaditud
domeenimoodulitesse (mitte enam legacy-moodulisse); kasutamata call()
eemaldatud.

862 testi läbivad, compileall puhas.
